{"schema":"decodex/commit/1","summary":"Split MCP HTTP handler modules","authority":"XY-1115"}
