+++ apps/decodex/src/mcp/http/handler/server.rs
@@ -1,0 +1,91 @@
+use std::{
+	io::{ErrorKind, Read as _, Write as _},
+	net::{TcpListener, TcpStream},
+};
+
+use crate::{
+	mcp::{
+		McpCapabilityProfile, McpContext, McpServer, McpTransport,
+		http::{
+			MCP_HTTP_MAX_REQUEST_BYTES, MCP_HTTP_READ_TIMEOUT,
+			auth::McpHttpAuthorization,
+			handler::{request::McpHttpHandler, sessions::McpHttpSessions},
+			message,
+		},
+	},
+	prelude::{Result, eyre},
+};
+
+pub(in crate::mcp) fn serve_streamable_http_with_profile(
+	listener: TcpListener,
+	context: McpContext,
+	capability_profile: McpCapabilityProfile,
+	allowed_origins: Vec<String>,
+	authorization: McpHttpAuthorization,
+) -> Result<()> {
+	let mut handler = McpHttpHandler {
+		server: McpServer { context, capability_profile, transport: McpTransport::StreamableHttp },
+		sessions: McpHttpSessions::default(),
+		allowed_origins,
+		listen_address: listener.local_addr().map(|address| address.to_string()).ok(),
+		authorization,
+	};
+
+	for stream in listener.incoming() {
+		match stream {
+			Ok(mut stream) => {
+				if let Err(error) = handle_mcp_http_stream(&mut stream, &mut handler) {
+					tracing::warn!(?error, "Decodex MCP Streamable HTTP request failed.");
+				}
+			},
+			Err(error) if error.kind() == ErrorKind::Interrupted => continue,
+			Err(error) => return Err(error.into()),
+		}
+	}
+
+	Ok(())
+}
+
+fn handle_mcp_http_stream(stream: &mut TcpStream, handler: &mut McpHttpHandler) -> Result<()> {
+	stream.set_read_timeout(Some(MCP_HTTP_READ_TIMEOUT))?;
+
+	let request = read_mcp_http_request(stream)?;
+	let response = handler.handle_request_bytes(&request)?;
+
+	stream.write_all(&response)?;
+	stream.flush()?;
+
+	Ok(())
+}
+
+fn read_mcp_http_request(stream: &mut TcpStream) -> Result<Vec<u8>> {
+	let mut buffer = Vec::new();
+	let mut scratch = [0_u8; 1_024];
+	let mut expected_len = None;
+
+	loop {
+		let read = stream.read(&mut scratch)?;
+
+		if read == 0 {
+			break;
+		}
+
+		buffer.extend_from_slice(&scratch[..read]);
+
+		if buffer.len() > MCP_HTTP_MAX_REQUEST_BYTES {
+			eyre::bail!("MCP HTTP request exceeded {MCP_HTTP_MAX_REQUEST_BYTES} bytes.");
+		}
+		if expected_len.is_none()
+			&& let Some(header_end) = message::http_header_end(&buffer)
+		{
+			let content_length = message::http_content_length(&buffer[..header_end])?;
+
+			expected_len = Some(header_end + 4 + content_length);
+		}
+		if expected_len.is_some_and(|length| buffer.len() >= length) {
+			break;
+		}
+	}
+
+	Ok(buffer)
+}
